Apply hobby policy

+++ app/Http/Controllers/HobbyController.php
@@ -7,6 +7,7 @@ use App\Tag;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
 use Intervention\Image\Facades\Image;
+use Illuminate\Support\Facades\Gate;
 
 class HobbyController extends Controller
 {
@@ -112,6 +113,8 @@ class HobbyController extends Controller
      */
     public function edit(Hobby $hobby)
     {
+        abort_unless(Gate::allows('update', $hobby), 403); // Không thể edit hobby của người khác
+
         return view('hobby.edit')->with([
             'hobby' => $hobby,
             'message_success' => Session::get('message_success'),
@@ -128,6 +131,9 @@ class HobbyController extends Controller
      */
     public function update(Request $request, Hobby $hobby)
     {
+
+        abort_unless(Gate::allows('update', $hobby), 403); // Không thể update hobby của người khác
+
         $request->validate([
             'name' => 'required|min:3',
             'description' => 'required|min:5',
@@ -159,6 +165,9 @@ class HobbyController extends Controller
      */
     public function destroy(Hobby $hobby)
     {
+        abort_unless(Gate::allows('delete', $hobby), 403); // Không thể xoá hobby của người khác
+
+
         $oldName = $hobby->name;
         $hobby->delete();
         return $this->index()->with([
